// src/api/schema.d.ts
export interface components {
  schemas: {
// ...
    RunStatusResponse: {
      /** Cancel Requested */
      cancel_requested: boolean;
      /** Id */
      id: string;
      /** Mode */
      mode: string;
      /** Published Attempt Id */
      published_attempt_id: string | null;
      /** State */
      state: string;
      /** Version */
      version: number;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}